+++ Classes/DataProcessing/TableProcessor.php
@@ -16,6 +16,7 @@ use Brotkrueml\JobRouterData\Domain\Demand\TableDemandFactory;
 use Brotkrueml\JobRouterData\Domain\Repository\TableRepository;
 use Brotkrueml\JobRouterData\Exception\TableNotFoundException;
 use Brotkrueml\JobRouterData\Extension;
+use TYPO3\CMS\Core\Localization\Locale;
 use TYPO3\CMS\Core\Service\FlexFormService;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 use TYPO3\CMS\Frontend\ContentObject\DataProcessorInterface;
@@ -66,7 +67,9 @@ final class TableProcessor implements DataProcessorInterface
             $table = $this->tableRepository->findByUid($tableUid);
             $tableDemand = $this->tableDemandFactory->create($table);
 
-            $locale = $this->cObj->getRequest()->getAttribute('language')->getLocale();
+            // locale is casted to a string as in v12 the locale is an object with a __toString() method (in v11 it is a string)
+            // @todo Remove the cast when compatibility with TYPO3 v11 is dropped
+            $locale = (string)$this->cObj->getRequest()->getAttribute('language')->getLocale();
             $this->processedData['table'] = $tableDemand;
             $this->processedData['rows'] = $this->datasetConverter->convertFromJsonToArray($table, $locale);
             $this->addCacheTag($tableUid);