+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Models are no longer Extbase-based and are moved to the Domain/Entity namespace
 - Repositories are no longer Extbase-based
 - JobDataRepository is now injectable, method signatures have changed
+- ModifyColumnContentEvent->getLocale() might return a different locale in TYPO3 v12
 
 ### Fixed
 - Error in content element when table link is disabled or deleted